Clean URs on trigger deletion
