+++ pkg/background/update_request_controller.go
@@ -6,8 +6,8 @@ import (
 	"time"
 
 	"github.com/go-logr/logr"
-	kyvernov1 "github.com/kyverno/kyverno/api/kyverno/v1"
-	kyvernov1beta1 "github.com/kyverno/kyverno/api/kyverno/v1beta1"
+	kyverno "github.com/kyverno/kyverno/api/kyverno/v1"
+	urkyverno "github.com/kyverno/kyverno/api/kyverno/v1beta1"
 	"github.com/kyverno/kyverno/pkg/autogen"
 	common "github.com/kyverno/kyverno/pkg/background/common"
 	kyvernoclient "github.com/kyverno/kyverno/pkg/client/clientset/versioned"
@@ -69,7 +69,7 @@ type Controller struct {
 	Config config.Configuration
 }
 
-// NewController returns an instance of the Generate-Request Controller
+//NewController returns an instance of the Generate-Request Controller
 func NewController(
 	kubeClient kubernetes.Interface,
 	kyvernoClient kyvernoclient.Interface,
@@ -204,9 +204,7 @@ func (c *Controller) syncUpdateRequest(key string) error {
 
 	logger.V(3).Info("UR is marked successfully", "ur", ur.GetName(), "resourceVersion", ur.GetResourceVersion())
 	if err := c.ProcessUR(ur); err != nil {
-		logger.Info("failed to process the UR, triggering handle delete operation", "handler", ur.Status.Handler, "ur", ur.GetName(), "err", err)
-		err = c.HandleDeleteUR(*ur)
-		return fmt.Errorf("failed to process delete UR %s: %v", key, err)
+		return fmt.Errorf("failed to process UR %s: %v", key, err)
 	}
 
 	if err = c.UnmarkUR(ur); err != nil {
@@ -229,8 +227,8 @@ func (c *Controller) enqueueUpdateRequest(obj interface{}) {
 
 func (c *Controller) updatePolicy(old, cur interface{}) {
 	logger := c.log
-	oldP := old.(*kyvernov1.ClusterPolicy)
-	curP := cur.(*kyvernov1.ClusterPolicy)
+	oldP := old.(*kyverno.ClusterPolicy)
+	curP := cur.(*kyverno.ClusterPolicy)
 	if oldP.ResourceVersion == curP.ResourceVersion {
 		// Periodic resync will send update events for all known Namespace.
 		// Two different versions of the same replica set will always have different RVs.
@@ -268,7 +266,7 @@ func (c *Controller) updatePolicy(old, cur interface{}) {
 }
 
 func (c *Controller) addUR(obj interface{}) {
-	ur := obj.(*kyvernov1beta1.UpdateRequest)
+	ur := obj.(*urkyverno.UpdateRequest)
 	if ur.Status.Handler != "" {
 		return
 	}
@@ -276,8 +274,8 @@ func (c *Controller) addUR(obj interface{}) {
 }
 
 func (c *Controller) updateUR(old, cur interface{}) {
-	oldUr := old.(*kyvernov1beta1.UpdateRequest)
-	curUr := cur.(*kyvernov1beta1.UpdateRequest)
+	oldUr := old.(*urkyverno.UpdateRequest)
+	curUr := cur.(*urkyverno.UpdateRequest)
 	if oldUr.ResourceVersion == curUr.ResourceVersion {
 		// Periodic resync will send update events for all known Namespace.
 		// Two different versions of the same replica set will always have different RVs.
@@ -285,7 +283,7 @@ func (c *Controller) updateUR(old, cur interface{}) {
 	}
 	// only process the ones that are in "Pending"/"Completed" state
 	// if the UPDATE Request fails due to incorrect policy, it will be requeued during policy update
-	if curUr.Status.State != kyvernov1beta1.Pending {
+	if curUr.Status.State != urkyverno.Pending {
 		return
 	}
 
@@ -297,14 +295,14 @@ func (c *Controller) updateUR(old, cur interface{}) {
 
 func (c *Controller) deleteUR(obj interface{}) {
 	logger := c.log
-	ur, ok := obj.(*kyvernov1beta1.UpdateRequest)
+	ur, ok := obj.(*urkyverno.UpdateRequest)
 	if !ok {
 		tombstone, ok := obj.(cache.DeletedFinalStateUnknown)
 		if !ok {
 			logger.Info("Couldn't get object from tombstone", "obj", obj)
 			return
 		}
-		ur, ok = tombstone.Obj.(*kyvernov1beta1.UpdateRequest)
+		ur, ok = tombstone.Obj.(*urkyverno.UpdateRequest)
 		if !ok {
 			logger.Info("tombstone contained object that is not a Update Request CR", "obj", obj)
 			return
